Moved the tag classes back to the scanner client and added Task
The tag classes contain too much behavior that is specific to the
scanner client to put in the common package. However, we still needed a
container for the return value of the taskFinder, so we created a
"trackable" Task class and appended "Tag" to all of the original tag
classes.